fix(view-as): bound the test-assignment preview to its assignment (#892)
Staff previewing their own test submission enter a synthetic student identity:
the app fabricates `role: "student"` over the instructor's own profile while
keeping their `private_profile_id`. Postgres never sees that fabrication, so
every query keyed on a real `role = 'student'` enrollment returns nothing for
it. Following the student nav to the Assignments tab therefore reported "No
upcoming deadlines available" for a course whose assignments were all released,
which read as a half-built student view.

Bound the synthetic identity to the assignment it was entered from. Inside
that subtree the surfaces work, because they key on `profile_id` alone; outside
it, the cookie is inert and the real staff identity is returned, so the course
home, gradebook and `/manage/*` all behave as staff again. `isSelfViewAsScope`
is shared by the server (`getEffectiveCourseIdentity`, via a new `x-pathname`
request header, since layouts get no pathname and a soft navigation does not
re-run the reused course layout) and the client, which also clears the stale
cookie so returning to an assignment does not silently re-enter the preview.

The student assignments page becomes a server component that sends any
non-student identity to the management list. Without it, leaving the preview
just re-rendered the same empty page under the instructor identity, since the
dashboard RPC is empty for a staff profile either way.

For the course-wide student view instructors actually want, add a student
picker that switches into the read-only view of a real enrolled student — the
path that has an enrollment to draw on. It sits in the view-as banner during
the preview and on the Test Assignment page, and pages through the roster
because courses here exceed PostgREST's 1000-row cap.

No SQL is widened: `get_assignments_for_student_dashboard` and
`assignments_with_effective_due_dates` still require a real student row, which
is fine now that no staff identity reaches them.
