+++ app/course/[course_id]/manage/assignments/[assignment_id]/test/page.tsx
@@ -2,6 +2,7 @@
 
 import { CommitHistoryDialog } from "@/app/course/[course_id]/assignments/[assignment_id]/commitHistory";
 import CreateStudentReposButton from "@/app/course/[course_id]/assignments/createStudentReposButton";
+import { ViewAsStudentPicker } from "@/components/course/view-as-student-picker";
 import { TimeZoneAwareDate } from "@/components/TimeZoneAwareDate";
 import { ActiveSubmissionIcon } from "@/components/ui/active-submission-icon";
 import { useClassProfiles } from "@/hooks/useClassProfiles";
@@ -74,6 +75,14 @@ export default function TestAssignmentPage() {
         read-only banner, grade-release rules, rubric visibility, and hidden autograder-output behavior students see.
         Use the banner to exit student view and compare the instructor or grader view.
       </Text>
+      <Text fontSize="sm" color="fg.muted" mt={2}>
+        That preview covers this assignment only — it is your own staff account wearing a student&apos;s view, so
+        course-wide pages such as the Assignments list have no enrollment to draw on. To see the whole course as a
+        student sees it, view it as one of your enrolled students instead.
+      </Text>
+      <Box mt={2} mb={4} maxW="sm">
+        <ViewAsStudentPicker showLabel />
+      </Box>
       {/* {repository?.data.length ? (
         <CreateStudentReposButton syncAllPermissions />
       ): <></>} */}
